@@ -3,7 +3,7 @@ project(Interpreter VERSION 0.1)
  
 set(SRC main.cpp Commands.cpp Interpreter.cpp My_Stack.cpp)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 add_executable(Interpreter ${SRC})
