@@ -9,7 +9,7 @@ class Command {
 
 class Add: public Command {
     //Adding two values from the stack. 
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
@@ -17,7 +17,7 @@ class Add: public Command {
 
 class Sub: public Command {
     //Subtraction two values from the stack. 
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
@@ -25,7 +25,7 @@ class Sub: public Command {
 
 class Div: public Command {
     //Division two values from the stack. 
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack. Exception when dividing by zero
     void apply(My_Stack & _value) override;
@@ -33,7 +33,7 @@ class Div: public Command {
 
 class Mod: public Command {
     //Taking the remainder of the division.
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack. Exception when dividing by zero
     void apply(My_Stack & _value) override;
@@ -41,7 +41,7 @@ class Mod: public Command {
 
 class Mul: public Command {
     //Multiplication two values from the stack. 
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
@@ -49,7 +49,7 @@ class Mul: public Command {
 
 class More: public Command {
     //Compares two values from the stack(>).
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
@@ -57,7 +57,7 @@ class More: public Command {
 
 class Less: public Command {
     //Compares two values from the stack(<).
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
@@ -65,45 +65,45 @@ class Less: public Command {
 
 class Equal: public Command {
     //Compares two values from the stack(==).
-    //The top of the stack is used as the right operand, and the second number on the stack is used as the left operand.
+    //The peek of the stack is used as the right operand, and the second number on the stack is used as the left operand.
     //The result of the operation is put on the stack.
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
 };
 
 class Write: public Command {
-    //Printing the top of the stack
+    //Printing the peek of the stack
     //Exception of interpreter_error if there are not enough elements on the stack
     void apply(My_Stack & _value) override;
 };
 
 class Dup: public Command {
-    //Copy the top of the stack and put it on the stack
+    //Copy the peek of the stack and put it on the stack
     void apply(My_Stack & _value) override;
 };
 
 class Drop: public Command {
-    //Remove the top number from the stack
+    //Remove the peek number from the stack
     void apply(My_Stack & _value) override;
 };
 
 class Swap: public Command {
-    //Swap the top and second numbers on the stack
+    //Swap the peek and second numbers on the stack
     void apply(My_Stack & _value) override;
 };
 
 class Rot: public Command {
-    //Cyclically shift the top three numbers on the stack
+    //Cyclically shift the peek three numbers on the stack
     void apply(My_Stack & _value) override;
 };
 
 class Over: public Command {
-    //Copy the second number and put a copy over the top one
+    //Copy the second number and put a copy over the peek one
     void apply(My_Stack & _value) override;
 };
 
 class Emit: public Command {
-    //Print the top number on the stack as an ascii code and remove it from the stack
+    //Print the peek number on the stack as an ascii code and remove it from the stack
     void apply(My_Stack & _value) override;
 };
 
